Merged PR 22020: engineering: Update `source` API
# 🔍 Description

Changes to API:
- Rename `Create` to `New`
- Make `OsImage` the default for `FileSystemSource`

This PR makes trident compatible with both the old and new APIs, with the exception being that `Create`/`New` is no longer the default and has to be specified.

Successful pr-e2e run:
https://dev.azure.com/mariner-org/ECF/_build/results?buildId=734157&view=results

# 🤔 Rationale

Why is this PR needed?

Create a bridge between the old and new APIs, where both are supported.

----
#### AI description  (iteration 1)
#### PR Classification
API change

#### PR Summary
This pull request updates the `source` API by removing partition image types from the `Source` enum and renaming `Create` to `New`.
- `trident_api/src/config/host/storage/filesystem.rs`: Added a temporary function to handle both new and old API, renamed `Create` to `New` in `FileSystemSource` enum.
- `docs/Reference/Host-Configuration/Sample-Host-Configuration.md` and other sample configuration files: Updated `source` field to use `new` type.
- `trident_api/src/samples/sample_hc.rs`: Replaced `FileSystemSource::Create` with `FileSystemSource::New`.
- `trident_api/schemas/host-config-schema.json`: Updated schema to reflect the new `source` field defaulting to `OS image` and renaming `create` to `new`.
- `src/engine/storage/filesystem.rs`: Updated logic to handle `FileSystemSource::New` instead of `Create`.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11013