Merged PR 22066: Fix error checking after new logic has been added
# 🔍 Description

Part of the loop-update.sh script is not using set -e for implicit error checking so that it can explicitly handle failures and report them to ADO. Recently, new commands were added without accounting for this behavior. The end result was that some failures were not detected early. This PR fixes the logic.

# 🤔 Rationale

Report failures on the stage command instead of reporting rollback later on.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request fixes error checking in the `loop-update.sh` script to ensure failures are correctly reported.
- `/scripts/loop-update/loop-update.sh`: Added explicit error handling for the `trident run` command and corrected the error checking mechanism to use the stored result.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11039