Merged PR 22071: Dont republish if already published instead of using jobattempts
# 🔍 Description

The previous logic for publishing Azure VM test images was using jobattempt counter, so that if a job had to be rerun, it would not conflict with a previously published image. That was however not working correctly, when the subsequent job also needed to be rerun, as their job attempts would not match. This was originally ok, as the publishing and consuming was in a single job. But with the split of the two, different logic had to be used. So instead, we will not republish if the image is already present.

# 🤔 Rationale

Fix handling of job reruns.

Related work items: #11039