Merged PR 22082: engineering: allow empty init on new api
# 🔍 Description

A slight modification to !21844 that allows for an old config to have _no_ image, and still receive an update.
This holds as long as the old spec is also using the new API.

----
#### AI description  (iteration 1)
#### PR Classification
API change

#### PR Summary
This pull request allows for an empty initialization on the new API by modifying the logic for OS image updates.
- `src/subsystems/storage/image.rs`: Adjusted the conditions to handle cases where the old spec did not request an OS image but the new spec does, allowing for a valid update if the old spec is not using the old API.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11055