Merged PR 22109: Add logic to test for kernel crashes
# 🔍 Description

This PR adds logic that will check for kernel dumps, that get produced when there is a kernel panic. If the dump if found, it will break the pipeline with the kernel panic information exported into the pipeline artifacts, so it can be investigated further. This expose any kernel crashes that we have been blind to, since kernel reboots post panic.

# 🤔 Rationale

The desire is to be able to triage any kernel panics.