Merged PR 21876: Bug: Unclear error when attempting to create directory for mount point
# 🔍 Description

Checks permissions before creating directories for internal mount points if the mount point does not exist. Now we:

- Issue a warning when attempting to create the directory.
- Verify that the (existent) parent directory is not read-only.
- Ensure the directory is not within a read-only mounted filesystem.

# 🤔 Rationale

If the mount point does not exist, we silently attempt to create it. This works most of the time but will fail if the directory would be within a read-only mounted filesystem. The error and the cause aren't clear enough by the Trident output.

# 📝 Checks

Example of new output:
![image (2).png](https://dev.azure.com/mariner-org/2311650c-e79e-4301-b4d2-96543fdd84ff/_apis/git/repositories/895b6b3d-5077-488a-8001-ab6b5a14c1a3/pullRequests/21876/attachments/image%20%282%29.png)

Related work items: #9579