Merged PR 22134: engineering: Don't run license check on pipeline critical path
This should hopefully cut 1-2 minutes of the PR E2E and CI pipeline runtimes. Downside is that it makes the non-E2E PR pipeline a bit longer

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup to optimize the pipeline process.

#### PR Summary
This pull request removes the license check from the pipeline's critical path and relocates it to a different stage.
- Removed license check steps from `/.pipelines/templates/stages/trident_rpms/check.yml`
- Added license check steps to `/.pipelines/templates/stages/validate_makefile/dev-build.yml`
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11117