Merged PR 22180: Fix using Azure VMs for scale test validation
# 🔍 Description

This PR appends the job index to the test RG that is created for the validation. This will allow each worker to use a different RG for creating VMs.

This PR also changes the prefix of the RG name to be more descriptive.

# 🤔 Rationale

Fixes trident-scale-official break as multiple workers are trying to use the same RG, resulting in conflicts.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request fixes the issue with using Azure VMs for scale test validation by updating the resource group naming convention.
- `.pipelines/templates/stages/testing_servicing/testing-template.yml`: Updated `TEST_RESOURCE_GROUP` to include `System.JobPositionInPhase` for better identification.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #10976