Merged PR 22193: bug: Don't return error if e2fsck is able to correct FS issues
I tested this locally by having Prism generate an image that was missing the /lost+found directory and confirmed that the bug happens on main but not with this commit.

#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request modifies the behavior of the `fix` function in `e2fsck.rs` to avoid returning an error if `e2fsck` is able to correct file system issues.
- `osutils/src/e2fsck.rs`: Updated `fix` function to check the output code of `e2fsck` and return `Ok(())` if the file system errors were corrected (output code 1).
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11194