Merged PR 22215: Always run cleanup
# 🔍 Description

To ensure that clean runs even if other tasks fail, moving cleanup to run as the last task, with condition always().

# 🤔 Rationale

We have been leaking RGs in Azure, as there were failures to create VMs, which then resulted in the fetch logs and cleanup task failure.

Always run cleanup

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request ensures that the cleanup script is always executed regardless of the job outcome.
- `.pipelines/templates/stages/testing_servicing/testing-template.yml`: Moved the `cleanup-vm.sh` script execution to always run, ensuring cleanup occurs even if the job fails.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #10976