Merged PR 22213: Update sku size now that we have a quota for intel skus
# 🔍 Description

This pipeline was using dadsv5 vm size to create azure VMs; this was because we had a quota for these due to using them for an oversized 1es pool. Since we switched the pool to intel ddsv5 vm size, we need to switch the Azure VMs as well.

# 🤔 Rationale

Fixes lack of quota issues when scheduling azure vms for testing.

Related work items: #11190