Merged PR 22234: Add support for testing on Azure VMs from before merging PRs
# 🔍 Description

This PR introduces a new pipeline, pr-e2e-azure. This PR is using the non-official template, and is configured to run against the official pool, that allows to test on Azure VMs. This new pipeline will be added as an optional pipeline to PRs and can be scheduled instead of pr-e2e when Azure VM changes are part of the PR. To prevent unexpected changes to the azure subscription, this pipeline is only allowed to be scheduled by build or project admins, so you would need your SC-alt or PIM elevate.

# 🤔 Rationale

Reduce the Azure VM related breaks post merge.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds support for testing on Azure VMs before merging pull requests.
- Added new pipeline configuration in `/.pipelines/trident-pr-e2e-official.yml` for Azure VM testing.
- Updated `/.pipelines/templates/pipeline-selector.yml` to include the new `pr-e2e-official` stage.
- Modified `/.pipelines/templates/e2e-template.yml` to support the new `pr-e2e-official` stage.
- Updated `/.pipelines/templates/stages/trident_rpms/trident-stage.yml` to include the new stage type.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #10976