Merged PR 22235: Enable boot diagnostics early only for Azure VMs and download the serial log...
# 🔍 Description

This PR will enable capturing serial console logs of Azure test VMs when there is a failure.

# 🤔 Rationale

When failure on Azure VM is hit, this will help with diagnosing the issue.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to enable boot diagnostics early for Azure VMs and download the serial log.

#### PR Summary
This pull request addresses the issue of enabling boot diagnostics early for Azure VMs and ensures the serial log is downloaded.
- `/scripts/loop-update/deploy-vm.sh`: Added logic to enable boot diagnostics early and wait for its availability.
- `/scripts/loop-update/fetch-logs.sh`: Added function to download Azure serial log and adjusted log download sequence.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #10976