Merged PR 22220: engineering: enable 'dev' base image concept
This was created because there was a change to containerd2 that the BMP team needed to validate. To do that, we needed both an image and the new containerd2 RPM from a specific pipeline run.

To that end, this PR allows specification of a pipeline run to pull dev images from.

If the specified pipeline contains a specific RPM change, downloading the RPMs is important for validation (for release/preview, the RPMs are published, but for a dev build they may not be).

requires:
* https://dev.azure.com/mariner-org/ECF/_git/platform-pipelines/pullrequest/22218
* https://dev.azure.com/mariner-org/ECF/_git/test-images/pullrequest/22219

validation: https://dev.azure.com/mariner-org/ECF/_build/results?buildId=746644&view=results

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces the concept of a 'dev' base image to the CI/CD pipeline.
- `/.pipelines/templates/e2e-template.yml`: Commented out host-based BM testing and adjusted container-based BM testing.
- `/.pipelines/templates/stages/common_tasks/find-base-image-version.yml`: Added logic to handle 'dev' base image version setting.
- `/.pipelines/templates/trident-platform-cicd-template.yml`: Changed `baseimgBuildType` to 'dev' for pre-stage.
- Added 'dev' option to `baseimgBuildType` parameter in multiple pipeline templates (`trident-container.yml`, `trident-testimg.yml`, `build-image.yml`, `vm-testing.yml`, `build-source.yml`, `release.yml`, `trident-stage.yml`, `trident-usb-iso.yml`).
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11231