Merged PR 22203: engineering: Use findmnt instead of HC to find mount points
# 🔍 Description

Eventually we want to remove `FileSystemType` from the API, so we will not be able to query the HC for the mount point and filesystem types of filesystems coming from an OS image. To reduce reliance on `FileSystemType`, this PR shifts the SELinux configuration step to use findmnt instead of data from the HC.

Note: I edited `findmnt-null-source.json` with `fstype` entries that I think are accurate, but they might not be 100%. This file is used in just one unit test to check that `source` field can be `null`.

----
#### AI description  (iteration 1)
#### PR Classification
Code improvement

#### PR Summary
This pull request replaces the use of Host Configuration (HC) with `findmnt` to find mount points in the SELinux configuration.
- `src/subsystems/selinux.rs`: Refactored `perform_relabel` function to use `findmnt` for retrieving mount points.
- `osutils/src/findmnt.rs`: Added `fstype` field to `MountpointMetadata` and updated `FINDMNT_COLUMNS` to include `fstype`.
- `trident_api/src/error.rs`: Added new error type `GetMountPointsInfo` and modified `GetRootMountPointInfo` error message.
- `osutils/src/test_files/findmnt-null-source.json`: Updated test data to include `fstype` entries.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11093