Merged PR 22362: Engineering: Remove newroot directory warning
# 🔍 Description

Add flag to skip warnings when creating newroot directory.

# 🤔 Rationale

Newroot directory is almost always created as it is a Trident concept which would rarely exist in an image.

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request refactors the handling of the newroot directory by removing the warning and improving the directory creation and verification logic.
- `src/engine/newroot.rs`: Added `create_mount_directory` and `verify_mount_directory` functions to streamline directory creation and verification. Refactored `prepare_mount_directory` and introduced `prepare_root_directory` to handle directory preparation more cleanly. Updated `get_new_root_path` to use the new `prepare_root_directory` function.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11444