Merged PR 22440: engineering: Create filesystems function for Engine Context + remove `Other` from API
# 🔍 Description

This PR creates a filesystems function, which should be used to get access to all filesystems from all sources. Once we remove filesystem type from the API for image-sourced filesystems, this will be useful so that we don't have to separately check the OS image and the host configuration for a full picture of all filesystems.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces a new function to manage filesystems within the Engine Context, addressing the removal of `FileSystemType` from the API.
- Added `src/engine/context/filesystem.rs` to define `FileSystemData` and implement filesystem-related methods in `EngineContext`.
- Updated `osutils/src/filesystems.rs` to include conversion from `FileSystemType` to `KnownFilesystemTypes`.
- Enhanced `trident_api/src/error.rs` with new error types for missing ESP and root filesystems.
- Moved `src/engine/context.rs` to `src/engine/context/mod.rs` without code changes.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11128