Merged PR 22507: feature: Support Slim Update COSI
# 🔍 Description

A "Slim Update COSI" file is a COSI with just the required images needed to drive an AB Update. Previously validation was very strict in only allowing _full_ COSI images, even on updates.

This PR changes validation a bit to be servicing-type-aware. On clean installs it demands all filesystems sourced from an image to be present, but on AB updates it only needs to find those placed on AB Volumes.

A full image is still accepted.

Also introduces a new internal param, `allowUnusedFilesystems`, to relax the check on unused filesystems the OS Image. This is needed by steamboat while we reach a final/long-term decision on the topic of empty FSs in the Prism definition.

# 🤔 Rationale

Smaller update images, like the ones that will be built once vm-servicing is a Prism feature.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces support for Slim Update COSI, enhancing the validation and generation of test contexts for different servicing types.
- `src/subsystems/storage/osimage.rs`: Added new test cases and modified existing ones to support A/B update servicing type and clean install servicing type.
- `src/subsystems/storage/osimage.rs`: Refactored `validate_filesystems` function to include checks for required filesystems based on servicing type.
- `trident_api/src/error.rs`: Added new error types and updated existing ones to provide more detailed error messages related to filesystem validation.
- `src/subsystems/storage/osimage.rs`: Introduced new helper functions `generate_test_context` and `generate_test_context_ab_update` for creating test contexts.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11598