Merged PR 22711: Bug: Enable metrics check for abupdate
# 🔍 Description

 Enable metrics check for abupdate.

# 🤔 Rationale

We weren't checking for A/B update metrics for A/B updates for rerun and memory constrained tests because we weren't enabled for A/B updates, but that was changed on: https://dev.azure.com/mariner-org/ECF/_git/trident/pullrequest/21455

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to enable metrics check for `ab-update`.

#### PR Summary
This pull request addresses a bug by modifying the conditions under which metrics are checked for `ab-update` operations.
- `.pipelines/templates/stages/testing_common/trident-metrics.yml`: Simplified the condition to always check for `ab-update` operations, regardless of the `tridentConfigurationName` value.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11869