Merged PR 22722: bug: Fix partitioning safety check fails when disk and kernel are out of sync
# 🔍 Description

the bug is insanely hard to reproduce because you need to get your system to a pretty broken state where the partition table and disk are out of sync, so I couldn't really think of a way to test this (i can't even repro manually...) but this should make sure we don't die when we see the issue.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses a bug where partitioning safety checks fail when the disk and kernel are out of sync.
- `osutils/src/lsblk.rs`: Introduced `get_opt` function to handle cases where no block device is found, and updated `get` function to use `get_opt`.
- `src/engine/storage/partitioning.rs`: Updated `partitioning_safety_check` to safely skip partitions not recognized by the kernel.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #8385