Merged PR 22795: engineering: Use storm for checking active volume in servicing tests
# 🔍 Description

Switch from parsing YAML in bash to using golang for it.

# 🤔 Rationale

More robust and better error messages. This will make offline-init work easier because we'll be able to see error messages if Trident fails to adopt the host.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces a new feature to use Storm for checking active volume in servicing tests.
- `check_ssh.go`: Added `checkActiveVolume` function to verify the active volume via SSH and updated the `CheckSshHelper` struct to include `ActiveVolume`.
- `common.sh`: Replaced the existing active volume check with a call to the new Storm helper.
- `testing-template.yml`: Added `storm-trident` to the list of artifacts and set executable permissions.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11315