Merged PR 22906: Set internal params in offline-init
# 🔍 Description

This adds functionality to set two internals params during offline init: uki support if the underlying Prism YAML indicates it, and a new "injectedHostStatus" flag that currently doesn't do anything but indicates that the HS was constructed outside of Trident.

----
#### AI description  (iteration 1)
#### PR Classification
New feature implementation.

#### PR Summary
This pull request implements internal parameter settings in the offline initialization process, contributing to the secure servicing of VMs running Azure Linux. It addresses the implementation of official support as outlined in the linked work item.
- In `src/offline_init.rs`, added logic to set `ENABLE_UKI_SUPPORT` based on configuration history and inject `injectedHostStatus` into host status.
- Introduced `PrismOs` and updated `PrismHistoryConfig` to include `os` and `preview_features`.
- Made `InternalParams` struct public in `trident_api/src/config/host/internal_params.rs`.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11315