Merged PR 22791: Bug: Change EFI Boot file name according to system's architecture
# 🔍 Description

- Selects the name of the EFI boot executable based on the architecture.
- Removes the other unsupported architectures.

# 🤔 Rationale

When testing on ARM64 during clean install, Trident couldn't find the bootloader in the specified path as we were using a name specific to the Amd64 architecture: `bootx64.efi`. The EFI Boot file was created with the correct name for the architecture: `bootaa64.efi`

# Checks

Before the change, Trident errors out on ARM64 with:
```
   0: Bootloader path '/EFI/AZLA/bootx64.efi' does not exist
```
As the EFI executable is named `bootaa64.efi`. After the change when locally testing on ARM64 the clean install passes that point and reaches reboot.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses a clean install error on ARM64 systems by dynamically selecting the correct EFI boot file name based on the system's architecture.
- `src/engine/bootentries.rs`: Added functions to determine the correct EFI boot file name based on architecture and updated references to use this dynamic name.
- `trident_api/src/error.rs`: Added a new error type for failures in retrieving the correct EFI boot name.
- `src/engine/boot/esp.rs`: Made the function `current_arch_efi_str` public to be used in determining the EFI boot file name.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #11919