Merged PR 22927: Engineering: Trigger trident-ci for changes in test-images repo
# 🔍 Description

Trigger trident-ci for changes in test-images repo

# 🤔 Rationale

This will help to catch errors caused on test-images sooner.

----
#### AI description  (iteration 1)
#### PR Classification
Pipeline configuration update to trigger CI for changes in the test-images repository.

#### PR Summary
This pull request updates the CI pipeline configuration to automatically trigger builds when changes are made to the `main` branch of the `test-images-Official` repository.
- Changes in `/.pipelines/trident-ci.yml`: Added a new pipeline trigger for the `test-images-Official` repository on the `main` branch.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #12032