Merged PR 22984: doc: rfc: COSI 1.1 Revision
# 🔍 Description

COSI 1.1 revision to add bootloader metadata and resolve some things that annoy me.

----
#### AI description  (iteration 1)
#### PR Classification
Documentation update for the revision of the Composable OS Image (COSI) specification to version 1.1.

#### PR Summary
This pull request revises the COSI specification document to include new fields and objects related to bootloader configuration and filesystem metadata. It introduces version 1.1 of the specification with detailed descriptions of new components.
- `Composable-OS-Image.md`: Added `Bootloader` object and related enums and objects for bootloader configuration, including `BootloaderType`, `SystemDBoot`, and `SystemDBootEntry`.
- `Composable-OS-Image.md`: Updated metadata schema to include `bootloader` and `filesystems` fields, with `images` now deprecated as an alias for `filesystems`.
- `Composable-OS-Image.md`: Revised `VerityConfig` and `ImageFile` objects to include versioning information.
- `Composable-OS-Image.md`: Updated `OsPackage` object to include versioning information for fields.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #12113