Merged PR 23274: engineering: Enable UKI + Encryption - Internal param to override sealing PCR
# 🔍 Description

UKIs seem to be changing the value of PCR 7, which we currently use for sealing encryption keys. This PR adds an internal param to override the default and updates config with UKI+encryption to use PCR 0 intead while we work on the overall encryption strategy.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request implements a new feature that enables overriding the default sealing PCR by introducing a new internal parameter.

#### PR Summary
The changes allow encryption provisioning to use a custom PCR binding instead of the default PCR7 by reading the `overrideEncryptionPcrs` internal parameter. This update increases flexibility for UKI+Encryption and includes corresponding adjustments in test configurations.
- **`src/engine/storage/encryption.rs`**: Added logic to parse and apply the new `overrideEncryptionPcrs` parameter and adjusted the encryption workflow to pass custom PCR flags.
- **`trident_api/src/constants.rs`**: Introduced the constant `OVERRIDE_ENCRYPTION_PCRS` for internal parameter support.
- **e2e test configurations**: Updated various `trident-config.yaml` and `target-configurations.yaml` files to include the new parameter and re-enable related test cases.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #12276